requirements.txt
fastapi
uvicorn
pymysql
pymongo
python-dotenv
cryptography
requests